chore(deps): upgrade tree-sitter 0.25→0.26
Bump tree-sitter parent crate 0.25 → 0.26.

Breaking API change: `Node::child` now takes `u32` instead of `usize`.
Adapt call-sites in `src/analysis/query/` to cast loop indices.

Add adversarial unit tests for parser robustness (empty, malformed,
binary garbage, long line, deep nesting inputs).

Closes #138.

+++ src/analysis/query/calls.rs
@@ -35,7 +35,7 @@ fn collect_rust_calls(
         }
     }
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             collect_rust_calls(source, child, target, results);
         }
     }
@@ -59,7 +59,7 @@ fn collect_js_calls(
         }
     }
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             collect_js_calls(source, child, target, results);
         }
     }
@@ -83,7 +83,7 @@ fn collect_python_calls(
         }
     }
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             collect_python_calls(source, child, target, results);
         }
     }
@@ -107,7 +107,7 @@ fn collect_go_calls(
         }
     }
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             collect_go_calls(source, child, target, results);
         }
     }
